plaani-service/src/main/java/plaani/service/validation/PasswordValidator.java
package plaani.service.validation;

import javax.validation.ConstraintValidator;
import javax.validation.ConstraintValidatorContext;

public class PasswordValidator implements ConstraintValidator<Password, String> {
    @Override
    public void initialize(Password constraintAnnotation) {

    }

    @Override
    public boolean isValid(String password, ConstraintValidatorContext context) {
        if (password == null) {
            return true;
        }
        return password.matches(".*[a-z]+.*")
                && password.matches(".*[A-Z]+.*")
                && password.matches(".*[0-9]+.*");
    }
}
